WIP Updated UI shared component EServiceContentInfo
