+++ src/components/Shared/EServiceContentInfo.tsx
@@ -1,42 +1,180 @@
-import { Box, Chip, Grid, Typography } from '@mui/material'
+import { Chip, Divider, Grid, Stack, Typography } from '@mui/material'
 import { AxiosResponse } from 'axios'
-import has from 'lodash/has'
-import React, { FunctionComponent } from 'react'
+import React, { FunctionComponent, useMemo, useState } from 'react'
 import {
-  AttributeKey,
-  BackendAttribute,
   EServiceDescriptorRead,
   EServiceDocumentRead,
   EServiceReadType,
-  GroupBackendAttribute,
-  SingleBackendAttribute,
+  FrontendAttributes,
 } from '../../../types'
 import { RunActionOutput, useFeedback } from '../../hooks/useFeedback'
 import { useRoute } from '../../hooks/useRoute'
 import { secondsToHoursMinutes } from '../../lib/format-utils'
 import { downloadFile } from '../../lib/file-utils'
-import { buildDynamicPath } from '../../lib/router-utils'
-import { DescriptionBlock } from '../DescriptionBlock'
-import { ResourceList } from './ResourceList'
-import { StyledAccordion } from './StyledAccordion'
+import { buildDynamicRoute } from '../../lib/router-utils'
 import { StyledLink } from './StyledLink'
-import sortBy from 'lodash/sortBy'
 import { formatThousands } from '../../lib/format-utils'
 import { useTranslation } from 'react-i18next'
-import { getDownloadDocumentName } from '../../lib/eservice-utils'
+import { getDownloadDocumentName, getLatestActiveVersion } from '../../lib/eservice-utils'
+import { remapBackendAttributesToFrontend } from '../../lib/attributes'
+import { StyledButton } from './StyledButton'
+import { InformationRow } from '../InformationRow'
+import StyledSection from './StyledSection'
+import { useHistory } from 'react-router-dom'
+import { StyledForm } from './StyledForm'
+import { StyledInputControlledSelect } from './StyledInputControlledSelect'
+import { AttributeSection } from '../AttributeSection'
+import { CHIP_COLORS_E_SERVICE, eServiceHelpLink, verifyVoucherHelpLink } from '../../lib/constants'
+import { WELL_KNOWN_URL } from '../../lib/env'
+import { AttachFile as AttachFileIcon, Launch as LaunchIcon } from '@mui/icons-material'
+import { ButtonNaked } from '@pagopa/mui-italia'
 
 type EServiceContentInfoProps = {
   data: EServiceReadType
+  descriptorId: string
 }
 
-export const EServiceContentInfo: FunctionComponent<EServiceContentInfoProps> = ({ data }) => {
-  const { t } = useTranslation(['eservice', 'attribute', 'common'])
-  const { runAction } = useFeedback()
-  const { routes } = useRoute()
+export const EServiceContentInfo: FunctionComponent<EServiceContentInfoProps> = ({
+  data,
+  descriptorId,
+}) => {
+  const frontendAttributes = useMemo(() => {
+    if (!data.attributes) return
+    return remapBackendAttributesToFrontend(data.attributes)
+  }, [data])
+
+  const activeDescriptor = getLatestActiveVersion(data)
+  const isCurrentVersion = activeDescriptor?.id === descriptorId
+
+  return (
+    <React.Fragment>
+      <Grid spacing={2} container>
+        <Grid item xs={7}>
+          <GeneralInfoSection data={data} />
+          <VersionInfoSection data={data} isCurrentVersion={isCurrentVersion} />
+        </Grid>
+        <Grid item xs={5}>
+          <DownloadSection data={data} />
+          <VoucherVerificationSection />
+        </Grid>
+      </Grid>
+
+      {frontendAttributes && <AttributesSection frontendAttributes={frontendAttributes} />}
+
+      <Grid spacing={2} container>
+        <Grid item xs={6}>
+          <VersionHistorySection
+            viewindDescriptorId={descriptorId!}
+            eserviceId={data.id}
+            descriptors={data.descriptors}
+          />
+        </Grid>
+      </Grid>
+    </React.Fragment>
+  )
+}
+
+function GeneralInfoSection({ data }: { data: EServiceReadType }) {
+  const { t } = useTranslation('eservice', {
+    keyPrefix: 'contentInfo.sections.generalInformations',
+  })
+  return (
+    <StyledSection>
+      <StyledSection.Title>{t('title')}</StyledSection.Title>
+      <StyledSection.Content>
+        <InformationRow label={t('technology')}>{data.technology}</InformationRow>
+      </StyledSection.Content>
+    </StyledSection>
+  )
+}
+
+function VersionInfoSection({
+  data,
+  isCurrentVersion,
+}: {
+  data: EServiceReadType
+  isCurrentVersion: boolean
+}) {
+  const { t } = useTranslation('eservice', {
+    keyPrefix: 'contentInfo.sections.versionInformations',
+  })
+  const { t: tCommon } = useTranslation('common')
+
   const activeDescriptor = data.activeDescriptor as EServiceDescriptorRead
 
-  // Get all documents actual URL
-  const wrapDownloadDocument = (document: EServiceDocumentRead) => async () => {
+  const getFormattedVoucherLifespan = () => {
+    const { hours, minutes } = secondsToHoursMinutes(activeDescriptor.voucherLifespan)
+
+    const minutesLabel = tCommon('time.minute', { count: minutes })
+    const hoursLabel = tCommon('time.hour', { count: hours })
+    const and = tCommon('conjunctions.and')
+
+    if (hours === 0) {
+      return `${minutes} ${minutesLabel}`
+    }
+
+    if (minutes === 0) {
+      return `${hours} ${hoursLabel}`
+    }
+
+    return `${hours} ${hoursLabel} ${and} ${minutes} ${minutesLabel}`
+  }
+
+  return (
+    <StyledSection>
+      <StyledSection.Title>{t('title')}</StyledSection.Title>
+      <StyledSection.Content>
+        <Stack spacing={2}>
+          <InformationRow label={t('actualVersion')}>
+            <Stack spacing={1} direction="row" alignItems="center">
+              <Typography component="span">{activeDescriptor.version}</Typography>
+              {isCurrentVersion && <Chip size="small" label="versione corrente" color="primary" />}
+            </Stack>
+          </InformationRow>
+          <InformationRow label={t('versionStatus')}>
+            <Chip
+              size="small"
+              label={tCommon(`status.eservice.${activeDescriptor.state || 'DRAFT'}`)}
+              color={CHIP_COLORS_E_SERVICE[activeDescriptor.state]}
+            />
+          </InformationRow>
+          <InformationRow label={t('description')}>{activeDescriptor?.description}</InformationRow>
+          <InformationRow label={t('audience')} labelDescription={t('audienceDescription')}>
+            {activeDescriptor.audience.join(', ')}
+          </InformationRow>
+          <InformationRow label={t('voucherLifespan')}>
+            {getFormattedVoucherLifespan()}
+          </InformationRow>
+          <InformationRow label={t('dailyCallsPerConsumer')}>
+            {formatThousands(activeDescriptor.dailyCallsPerConsumer)} {t('callsPerDay')}
+          </InformationRow>
+          <InformationRow label={t('dailyCallsTotal')}>
+            {formatThousands(activeDescriptor.dailyCallsTotal)} {t('callsPerDay')}
+          </InformationRow>
+
+          <Divider />
+
+          <Typography variant="body2">
+            {t('doubtsQuestion')}{' '}
+            <StyledLink href={eServiceHelpLink} target="_blank">
+              {t('doubtsLink')}
+            </StyledLink>
+          </Typography>
+        </Stack>
+      </StyledSection.Content>
+    </StyledSection>
+  )
+}
+
+function DownloadSection({ data }: { data: EServiceReadType }) {
+  const { t } = useTranslation('eservice', {
+    keyPrefix: 'contentInfo.sections.download',
+  })
+  const { runAction } = useFeedback()
+
+  const activeDescriptor = data.activeDescriptor as EServiceDescriptorRead
+
+  const handleDownloadDocument = async (document: EServiceDocumentRead) => {
     const { response, outcome } = (await runAction(
       {
         path: {
@@ -58,213 +196,164 @@ export const EServiceContentInfo: FunctionComponent<EServiceContentInfoProps> = 
     }
   }
 
-  const getVerificationRequiredStringMaybe = (
-    explicitAttributeVerification: boolean,
-    attributeKey: AttributeKey
-  ) => {
-    return explicitAttributeVerification && attributeKey === 'verified'
-      ? ` (${t('contentInfo.verificationRequired')})`
-      : ''
-  }
+  const docs = [...activeDescriptor.docs, activeDescriptor.interface]
 
-  const toAccordionEntries = (attributes: Array<BackendAttribute>, attributeKey: AttributeKey) => {
-    return attributes.map((attribute) => {
-      const isSingle = has(attribute, 'single')
+  return (
+    <StyledSection>
+      <StyledSection.Title>{t('title')}</StyledSection.Title>
+      <StyledSection.Content>
+        <Stack spacing={2} alignItems="start">
+          {docs.map((doc) => (
+            <Stack key={doc.id} spacing={2}>
+              <ButtonNaked
+                onClick={handleDownloadDocument.bind(null, doc)}
+                size="large"
+                color="primary"
+                startIcon={<AttachFileIcon />}
+              >
+                {doc.prettyName}
+              </ButtonNaked>
+              {/* TEMP BACKEND - Size data doesn't come from backend (yet) */}
+              {/* <Typography fontWeight={600} sx={{ marginLeft: '30px' }}>
+                {(doc.size / 1024).toFixed(2)}&nbsp;KB
+              </Typography> */}
+            </Stack>
+          ))}
+        </Stack>
+      </StyledSection.Content>
+    </StyledSection>
+  )
+}
 
-      const labels = isSingle
-        ? [(attribute as SingleBackendAttribute).single]
-        : (attribute as GroupBackendAttribute).group
+function VoucherVerificationSection() {
+  const { t } = useTranslation(['eservice', 'common'], {
+    keyPrefix: 'contentInfo.sections.voucherVerification',
+  })
 
-      let summary: string | JSX.Element
-      let details: string | JSX.Element = ''
-      if (labels.length === 1) {
-        const { name, description, explicitAttributeVerification } = labels[0]
-        summary = `${name} ${getVerificationRequiredStringMaybe(
-          explicitAttributeVerification,
-          attributeKey
-        )}`
-
-        details = description
-      } else {
-        const summaryComponent = labels.map(({ name }, i) => {
-          if (i < labels.length - 1) {
-            return (
-              <React.Fragment key={i}>
-                {name}{' '}
-                <Typography component="span" fontWeight={600}>
-                  {t('contentInfo.groupOr')}
-                </Typography>{' '}
-              </React.Fragment>
-            )
-          }
-
-          return name
-        })
-
-        summary = (
-          <React.Fragment>
-            {summaryComponent}{' '}
-            {getVerificationRequiredStringMaybe(
-              labels[0].explicitAttributeVerification,
-              attributeKey
-            )}
-          </React.Fragment>
-        )
-
-        details = (
-          <React.Fragment>
-            {labels.map((label, i) => {
-              return (
-                <Box sx={{ mb: i !== labels.length - 1 ? 2 : 0 }} key={i}>
-                  <Typography component="span" sx={{ fontWeight: 700 }}>
-                    {label.name}
-                  </Typography>
-                  : {label.description}
-                </Box>
-              )
-            })}
-          </React.Fragment>
-        )
-      }
-
-      return { summary, details }
-    })
-  }
-
-  const getFormattedVoucherLifespan = () => {
-    const { hours, minutes } = secondsToHoursMinutes(activeDescriptor.voucherLifespan)
-
-    const minutesLabel = t('time.minute', { count: minutes, ns: 'common' })
-    const hoursLabel = t('time.hour', { count: hours, ns: 'common' })
-    const and = t('conjunctions.and', { ns: 'common' })
-
-    if (hours === 0) {
-      return `${minutes} ${minutesLabel}`
-    }
-
-    if (minutes === 0) {
-      return `${hours} ${hoursLabel}`
-    }
-
-    return `${hours} ${hoursLabel} ${and} ${minutes} ${minutesLabel}`
+  function VoucherLink({ label, href }: { label: string; href: string }) {
+    return (
+      <StyledLink
+        component="a"
+        href={href}
+        target="_blank"
+        variant="body2"
+        underline="hover"
+        sx={{ display: 'flex', alignItems: 'center', cursor: 'pointer' }}
+      >
+        <LaunchIcon sx={{ mr: 1 }} /> {label}
+      </StyledLink>
+    )
   }
 
   return (
-    <React.Fragment>
-      <DescriptionBlock label={t('contentInfo.provider')} sx={{ mt: 0 }}>
-        <Typography component="span">{data.producer.name}</Typography>
-      </DescriptionBlock>
+    <StyledSection>
+      <StyledSection.Title>{t('title')}</StyledSection.Title>
+      <StyledSection.Content>
+        <Typography>{t('description')}</Typography>
 
-      <DescriptionBlock label={t('contentInfo.version')}>
-        <Typography component="span">{activeDescriptor.version}</Typography>
-      </DescriptionBlock>
+        <Stack sx={{ mt: 2 }} spacing={2}>
+          <VoucherLink label={t('howLink')} href={verifyVoucherHelpLink} />
+          <VoucherLink label={t('wellKnownLink')} href={WELL_KNOWN_URL} />
+        </Stack>
+      </StyledSection.Content>
+    </StyledSection>
+  )
+}
 
-      <DescriptionBlock label={t('contentInfo.versionDescription')}>
-        <Typography component="span">{activeDescriptor.description}</Typography>
-      </DescriptionBlock>
+function AttributesSection({ frontendAttributes }: { frontendAttributes: FrontendAttributes }) {
+  const { t } = useTranslation('eservice', {
+    keyPrefix: 'contentInfo.sections.attributes',
+  })
 
-      <DescriptionBlock label={t('contentInfo.versionStatus')}>
-        <Typography component="span">
-          {t(`status.eservice.${activeDescriptor.state}`, { ns: 'common' })}
-        </Typography>
-      </DescriptionBlock>
+  return (
+    <>
+      <AttributeSection
+        attributeKey="certified"
+        description={t('certified.description')}
+        attributesSubtitle={t('subtitle')}
+        attributes={frontendAttributes.certified}
+        readOnly
+      />
+      <AttributeSection
+        attributeKey="verified"
+        description={t('verified.description')}
+        attributesSubtitle={t('subtitle')}
+        attributes={frontendAttributes.verified}
+        readOnly
+      />
+      <AttributeSection
+        attributeKey="declared"
+        description={t('declared.description')}
+        attributesSubtitle={t('subtitle')}
+        attributes={frontendAttributes.declared}
+        readOnly
+      />
+    </>
+  )
+}
 
-      <DescriptionBlock label={t('contentInfo.audience')}>
-        <Typography component="span">{activeDescriptor.audience.join(', ')}</Typography>
-      </DescriptionBlock>
+function VersionHistorySection({
+  viewindDescriptorId,
+  eserviceId,
+  descriptors,
+}: {
+  viewindDescriptorId: string
+  eserviceId: string
+  descriptors: EServiceDescriptorRead[]
+}) {
+  const { t } = useTranslation(['eservice', 'common'], {
+    keyPrefix: 'contentInfo.sections.versionHistory',
+  })
 
-      <DescriptionBlock label={t('contentInfo.technology')}>
-        <Typography component="span">{data.technology}</Typography>
-      </DescriptionBlock>
+  const history = useHistory()
+  const { routes } = useRoute()
+  const [selectedDescriptorId, setSelectedDescriptorId] = useState(viewindDescriptorId)
 
-      <DescriptionBlock label={t('contentInfo.voucherLifespan')}>
-        <Typography component="span">{getFormattedVoucherLifespan()}</Typography>
-      </DescriptionBlock>
+  const handleVersionChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setSelectedDescriptorId(e.target.value)
+  }
 
-      <DescriptionBlock label={t('contentInfo.dailyCallsPerConsumer')}>
-        <Typography component="span">
-          {formatThousands(activeDescriptor.dailyCallsPerConsumer)} {t('contentInfo.callsPerDay')}
-        </Typography>
-      </DescriptionBlock>
+  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault()
+    history.push(
+      buildDynamicRoute(routes.PROVIDE_ESERVICE_MANAGE, {
+        eserviceId,
+        descriptorId: selectedDescriptorId,
+      }).PATH
+    )
+  }
 
-      <DescriptionBlock label={t('contentInfo.dailyCallsTotal')}>
-        <Typography component="span">
-          {formatThousands(activeDescriptor.dailyCallsTotal)} {t('contentInfo.callsPerDay')}
-        </Typography>
-      </DescriptionBlock>
+  const shouldNotRender = descriptors.length <= 1
 
-      {(Object.keys(data.attributes) as Array<AttributeKey>).map((key, i) => (
-        <DescriptionBlock
-          key={i}
-          label={`${t('contentInfo.attributes')} ${t(`type.${key}`, {
-            count: 2,
-            ns: 'attribute',
-          })}`}
-        >
-          {data.attributes[key].length > 0 ? (
-            <Grid container>
-              <Grid item xs={8}>
-                <StyledAccordion entries={toAccordionEntries(data.attributes[key], key)} />
-              </Grid>
-            </Grid>
-          ) : (
-            <Typography component="span">{t('contentInfo.noAttributesLabel')}</Typography>
-          )}
-        </DescriptionBlock>
-      ))}
+  if (shouldNotRender) return null
 
-      <DescriptionBlock label={t('contentInfo.documentation')}>
-        <ResourceList
-          resources={[
-            // TEMP PIN-1095 and PIN-1105
-            // {
-            //   label: 'Richiesta di fruizione',
-            //   onClick: () => {
-            //     console.log('download richiesta di fruizione')
-            //   },
-            // },
-            {
-              label: activeDescriptor.interface.prettyName,
-              onClick: wrapDownloadDocument(activeDescriptor.interface),
-            },
-            ...activeDescriptor.docs.map((d) => ({
-              label: d.prettyName,
-              onClick: wrapDownloadDocument(d),
-            })),
-          ]}
-        />
-      </DescriptionBlock>
+  const descriptorsOptions = descriptors.map((descriptor) => ({
+    label: t('historyField.option', { version: descriptor.version }),
+    value: descriptor.id,
+  }))
 
-      {Boolean(data.descriptors.length > 0) && (
-        <DescriptionBlock label={t('contentInfo.versionHistory')} sx={{ mb: 0 }}>
-          {sortBy(data.descriptors, 'version').map((d, i) => {
-            const state = t(`status.eservice.${d.state}`, { ns: 'common' })
-
-            return (
-              <Box key={i} sx={{ pb: 1 }}>
-                {d.id !== data.activeDescriptor?.id ? (
-                  <Box>
-                    <StyledLink
-                      to={buildDynamicPath(routes.PROVIDE_ESERVICE_MANAGE.PATH, {
-                        eserviceId: data.id,
-                        descriptorId: d.id,
-                      })}
-                    >
-                      {t('contentInfo.version')} {d.version}
-                    </StyledLink>{' '}
-                    <Chip size="small" label={state} />
-                  </Box>
-                ) : (
-                  <Typography component="span">
-                    {t('contentInfo.version')} {d.version}{' '}
-                    <Chip size="small" label={state} color="primary" />
-                  </Typography>
-                )}
-              </Box>
-            )
-          })}
-        </DescriptionBlock>
-      )}
-    </React.Fragment>
+  return (
+    <StyledSection>
+      <StyledSection.Title>{t('title')}</StyledSection.Title>
+      <StyledSection.Content>
+        <StyledForm onSubmit={handleSubmit}>
+          <InformationRow label={t('historyField.title')}>
+            <StyledInputControlledSelect
+              label={t('historyField.label')}
+              sx={{ width: '100%' }}
+              emptyLabel=""
+              name="eServiceHistorySelection"
+              onChange={handleVersionChange}
+              value={selectedDescriptorId}
+              options={descriptorsOptions}
+              selectProps={{ MenuProps: { sx: { maxHeight: '160px' } } }}
+            />
+            <StyledButton sx={{ mt: 2 }} size="large" variant="outlined" type="submit">
+              {t('submitBtn')}
+            </StyledButton>
+          </InformationRow>
+        </StyledForm>
+      </StyledSection.Content>
+    </StyledSection>
   )
 }
